Add multi-edges between distant roots
Add 'multi edges' to convey dependency relationships between roots that are too distant to be otherwise linked based on the current extension depth.

Currently in display these are dashed instead of solid, and a tooltip shows the dependency chain that they represent.

The dependency chain is found by breadth-first search, the first shortest path found is used.