Improve selection handling for files with multiple types
Make the selection handling smarter about files with multiple types. The inner graph construction is already fairly robust in this regard.  The GraphUpdateManager and the view model now check if the current selection is a match for the active document, rather than always resetting the selection to the first type defined in the active document.

Instead of resetting the SelectedVertex every time the Graph changes in DependencyGraphLayout in order to force the selection highlights to be updated, we target the selection highlights specifically, to avoid pushing a spurious null selection to the view model.