Refactor Subgraph to use inclusion reason
Remove AdditionalNodes and PinnedNodes in favour of a generalised scheme whereby nodes are included in the subgraph under one or more nominated categories; these categories can be unapplied, and the node will be removed from the subgraph when it no longer has a reason to be there.

This alters the semantics of graph interactions somewhat, but most likely for the better. It paves the way to cleanly handle the introduction of new 'modes' with their own inclusion rules.